refactor(release): split plan_steps.rs god-file (#6181)
