+++ homeboy.json
@@ -794,7 +794,6 @@
         "structural::src/core/preview_ingress.rs::GodFile",
         "structural::src/core/preview_ingress.rs::HighItemCount",
         "structural::src/core/refactor/rename/mod.rs::GodFile",
-        "structural::src/core/release/plan_steps.rs::GodFile",
         "structural::src/core/runner/connection.rs::HighItemCount",
         "structural::src/core/runner/lab/agent_task_bridge.rs::HighItemCount",
         "structural::src/core/runner/lab/offload.rs::GodFile",
