+++ src/core/release/plan_steps/mod.rs
@@ -1,0 +1,11 @@
+mod builders;
+mod changelog;
+mod hints;
+mod preflight;
+mod release;
+
+pub(in crate::core::release) use preflight::build_preflight_steps;
+pub(in crate::core::release) use release::build_release_steps;
+
+#[cfg(test)]
+mod tests;
